email-saas/requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
joblib
scikit-learn
numpy
scipy
